MicroscopyLabPlanner - Update commit - Hour 2
Updated the progress bar and moved it to the bottom.Pressing on the stage number of the progress bar is an alternate method to move the the next or previous stage instead of clicking the button. The button is disabled if the checklist or dropdown menu is not filled in. Added an additional dropdown menu of cell density experiment is selected, it asks the user whether an led or halogen lamp microscope is used. placeholder checkbox is used for phase contrast and fluorescence, but I've written the full steps for when LED microscope is choosed for cell density experiment